+++ src/rx-query-helper.ts
@@ -10,7 +10,9 @@ import {
     clone,
     firstPropertyNameOfObject,
     toArray,
-    isMaybeReadonlyArray
+    isMaybeReadonlyArray,
+    parseRegex,
+    flatClone
 } from './plugins/utils';
 
 /**
@@ -22,8 +24,13 @@ export function normalizeMangoQuery<RxDocType>(
     mangoQuery: MangoQuery<RxDocType>
 ): FilledMangoQuery<RxDocType> {
     const primaryKey: string = getPrimaryFieldOfPrimaryKey(schema.primaryKey);
-    const normalizedMangoQuery: FilledMangoQuery<RxDocType> = clone(mangoQuery) as any;
+    mangoQuery = flatClone(mangoQuery);
 
+    // regex normalization must run before deep clone because deep clone cannot clone RegExp
+    if (mangoQuery.selector) {
+        mangoQuery.selector = normalizeQueryRegex(mangoQuery.selector);
+    }
+    const normalizedMangoQuery: FilledMangoQuery<RxDocType> = clone(mangoQuery) as any;
     if (typeof normalizedMangoQuery.skip !== 'number') {
         normalizedMangoQuery.skip = 0;
     }
@@ -149,4 +156,35 @@ export function normalizeMangoQuery<RxDocType>(
     }
 
     return normalizedMangoQuery;
+}
+
+/**
+ * @recursive
+ * @mutates the input so that we do not have to deep clone
+ */
+export function normalizeQueryRegex(
+    selector: any
+): any {
+    if (typeof selector !== 'object') {
+        return selector;
+    }
+
+    const keys = Object.keys(selector);
+    const ret: any = {};
+    keys.forEach(key => {
+        const value: any = selector[key];
+        if (
+            key === '$regex' &&
+            value instanceof RegExp
+        ) {
+            const parsed = parseRegex(value);
+            ret.$regex = parsed.pattern;
+            ret.$options = parsed.flags;
+        } else if (Array.isArray(value)) {
+            ret[key] = value.map(item => normalizeQueryRegex(item));
+        } else {
+            ret[key] = normalizeQueryRegex(value);
+        }
+    });
+    return ret;
 }